backend/requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
google-generativeai
